+++ Kassa/web-app/src/main/resources/db/migration/V20220118204000__create_event_table.sql
@@ -2,7 +2,7 @@ CREATE TABLE event
 (
     id          serial          NOT NULL,
     name        VARCHAR(25)     NOT NULL UNIQUE,
-    places      DOUBLE          NOT NULL,
+    places      Integer          NOT NULL,
     price       Integer         NOT NULL,
     date        timestamp       NOT NULL,
     PRIMARY KEY (id)
